+++ templates/react-component/src/index.js
@@ -1,9 +1,9 @@
-import React from 'react'
+import React, {Component} from 'react'
 
-export default React.createClass({
+export default class extends Component {
   render() {
     return <div>
       <h2>Welcome to React components</h2>
     </div>
   }
-})
+}