+++ tests/commands/build-test.js
@@ -32,7 +32,6 @@ describe('command: build', function() {
   }
 
   describe('building a React app', () => {
-    let builtAppSource
     let builtHTML
 
     before(done => {
@@ -42,7 +41,6 @@ describe('command: build', function() {
         process.chdir(path.join(tmpDir, 'test-app'))
         cli(['build'], (err) => {
           expect(err).toNotExist('No errors building a React app')
-          builtAppSource = fs.readFileSync(glob.sync('dist/app*.js')[0], 'utf8')
           builtHTML = fs.readFileSync('dist/index.html', 'utf8')
           done(err)
         })
@@ -64,9 +62,6 @@ describe('command: build', function() {
         'vendor.js',
         'vendor.js.map',
       ])
-    })
-    it('generates displayName for React.createClass calls in the build', () => {
-      expect(builtAppSource).toInclude('displayName:"App"')
     })
     it('injects the Webpack manifest into generated HTML', () => {
       expect(builtHTML).toInclude('window.webpackJsonp')
@@ -124,10 +119,6 @@ describe('command: build', function() {
         /^demo\.\w{8}\.js\.map/,
         'index.html',
       ])
-    })
-    it('generates displayName for React.createClass calls in the demo build', () => {
-      expect(fs.readFileSync(glob.sync('demo/dist/demo*.js')[0], 'utf8'))
-        .toInclude('displayName:"Demo"')
     })
   })